requirements.txt
# Core DL stack
torch>=2.1.0
torchvision>=0.16.0

# Utilities
numpy>=1.24.0
Pillow>=9.4.0
tqdm>=4.65.0
scikit-learn>=1.2.0

# Optional (for saving reports)
matplotlib>=3.7.0
